requirements.txt
llama-index-core
llama-index-embeddings-ollama
streamlit
ollama
PyPDF2
